+++ Xplat/src/main/java/vazkii/botania/common/block/mana/DrumBlock.java
@@ -11,15 +11,15 @@ package vazkii.botania.common.block.mana;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.sounds.SoundSource;
-import net.minecraft.world.entity.Entity;
+import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.Shearable;
-import net.minecraft.world.entity.animal.Cow;
 import net.minecraft.world.entity.animal.MushroomCow;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.SuspiciousStewItem;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
@@ -35,14 +35,20 @@ import vazkii.botania.api.mana.ManaTrigger;
 import vazkii.botania.common.block.BotaniaWaterloggedBlock;
 import vazkii.botania.common.block.flower.functional.BergamuteBlockEntity;
 import vazkii.botania.common.handler.BotaniaSounds;
+import vazkii.botania.common.helper.EntityHelper;
 import vazkii.botania.common.item.BotaniaItems;
 import vazkii.botania.common.item.HornItem;
+import vazkii.botania.common.lib.BotaniaTags;
+import vazkii.botania.mixin.MushroomCowAccessor;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 public class DrumBlock extends BotaniaWaterloggedBlock {
+
+	public static final int MAX_NUM_SHEARED = 4;
+	public static final int GATHER_RANGE = 10;
 
 	public enum Variant {
 		WILD,
@@ -64,23 +70,85 @@ public class DrumBlock extends BotaniaWaterloggedBlock {
 		return SHAPE;
 	}
 
-	private static void convertNearby(Entity entity, Item from, Item to) {
-		Level world = entity.level();
-		List<ItemEntity> items = world.getEntitiesOfClass(ItemEntity.class, entity.getBoundingBox());
-		for (ItemEntity item : items) {
-			ItemStack itemstack = item.getItem();
-			if (!itemstack.isEmpty() && itemstack.is(from) && !world.isClientSide) {
-				while (itemstack.getCount() > 0) {
-					ItemEntity ent = entity.spawnAtLocation(new ItemStack(to), 1.0F);
-					ent.setDeltaMovement(ent.getDeltaMovement().add(
-							world.random.nextFloat() * 0.05F,
-							(world.random.nextFloat() - world.random.nextFloat()) * 0.1F,
-							(world.random.nextFloat() - world.random.nextFloat()) * 0.1F
-					));
-					itemstack.shrink(1);
-				}
-				item.discard();
+	public static void gatherProduce(Level world, BlockPos pos) {
+		List<Mob> mobs = world.getEntitiesOfClass(Mob.class, new AABB(pos.offset(-GATHER_RANGE, -GATHER_RANGE, -GATHER_RANGE), pos.offset(GATHER_RANGE + 1, GATHER_RANGE + 1, GATHER_RANGE + 1)),
+				mob -> mob.isAlive() && !BergamuteBlockEntity.isBergamuteNearby(world, mob.getX(), mob.getY(), mob.getZ()));
+		List<Shearable> shearables = new ArrayList<>();
+
+		for (Mob mob : mobs) {
+			if (mob.getType().is(BotaniaTags.Entities.DRUM_MILKABLE) && !mob.isBaby()) {
+				convertNearby(mob, Items.BUCKET, Items.MILK_BUCKET);
 			}
+			if (mob instanceof MushroomCow mooshroom && !mooshroom.isBaby()) {
+				if (mooshroom.getVariant() == MushroomCow.MushroomType.BROWN && ((MushroomCowAccessor) mooshroom).getEffect() != null) {
+					fillBowlSuspiciously(mooshroom);
+				}
+				convertNearby(mob, Items.BOWL, Items.MUSHROOM_STEW);
+			}
+			if (mob instanceof Shearable shearable && !mob.getType().is(BotaniaTags.Entities.DRUM_NO_SHEARING) && shearable.readyForShearing()) {
+				shearables.add(shearable);
+			}
+		}
+
+		Collections.shuffle(shearables);
+		int sheared = 0;
+
+		for (Shearable shearable : shearables) {
+			if (sheared > MAX_NUM_SHEARED) {
+				break;
+			}
+
+			shearable.shear(SoundSource.BLOCKS);
+			++sheared;
+		}
+	}
+
+	private static void convertNearby(Mob mob, Item from, Item to) {
+		Level world = mob.level();
+		List<ItemEntity> fromEntities = world.getEntitiesOfClass(ItemEntity.class, mob.getBoundingBox(),
+				itemEntity -> itemEntity.isAlive() && itemEntity.getItem().is(from));
+		for (ItemEntity fromEntity : fromEntities) {
+			ItemStack fromStack = fromEntity.getItem();
+			for (int i = fromStack.getCount(); i > 0; i--) {
+				spawnItem(mob, new ItemStack(to));
+			}
+			fromEntity.discard();
+		}
+	}
+
+	private static void spawnItem(Mob mob, ItemStack to) {
+		Level world = mob.level();
+		ItemEntity ent = mob.spawnAtLocation(to, 1.0F);
+		ent.setDeltaMovement(ent.getDeltaMovement().add(
+				world.random.nextFloat() * 0.05F,
+				(world.random.nextFloat() - world.random.nextFloat()) * 0.1F,
+				(world.random.nextFloat() - world.random.nextFloat()) * 0.1F
+		));
+	}
+
+	private static void fillBowlSuspiciously(MushroomCow mushroomCow) {
+		MushroomCowAccessor mushroomCowAccessor = (MushroomCowAccessor) mushroomCow;
+		MobEffect effect = mushroomCowAccessor.getEffect();
+		int effectDuration = mushroomCowAccessor.getEffectDuration();
+
+		Level world = mushroomCow.level();
+		List<ItemEntity> bowlItemEntities = world.getEntitiesOfClass(ItemEntity.class, mushroomCow.getBoundingBox(),
+				itemEntity -> itemEntity.getItem().is(Items.BOWL) && !itemEntity.getItem().isEmpty());
+		for (ItemEntity bowlItemEntity : bowlItemEntities) {
+			ItemStack bowlItem = bowlItemEntity.getItem();
+			ItemStack stewItem = new ItemStack(Items.SUSPICIOUS_STEW);
+			SuspiciousStewItem.saveMobEffect(stewItem, effect, effectDuration);
+			spawnItem(mushroomCow, stewItem);
+
+			EntityHelper.shrinkItem(bowlItemEntity);
+			if (bowlItem.getCount() == 0) {
+				bowlItemEntity.discard();
+			}
+
+			// only one suspicious stew per flower fed
+			mushroomCowAccessor.setEffect(null);
+			mushroomCowAccessor.setEffectDuration(0);
+			break;
 		}
 	}
 
@@ -104,40 +172,10 @@ public class DrumBlock extends BotaniaWaterloggedBlock {
 				world.addParticle(ParticleTypes.NOTE, pos.getX() + 0.5, pos.getY() + 1.2, pos.getZ() + 0.5D, 1.0 / 24.0, 0, 0);
 				return;
 			}
-			if (variant == Variant.WILD) {
-				HornItem.breakGrass(world, new ItemStack(BotaniaItems.grassHorn), pos, null);
-			} else if (variant == Variant.CANOPY) {
-				HornItem.breakGrass(world, new ItemStack(BotaniaItems.leavesHorn), pos, null);
-			} else {
-				int range = 10;
-				List<Mob> entities = world.getEntitiesOfClass(Mob.class, new AABB(pos.offset(-range, -range, -range), pos.offset(range + 1, range + 1, range + 1)), e -> !BergamuteBlockEntity.isBergamuteNearby(world, e.getX(), e.getY(), e.getZ()));
-				List<Mob> shearables = new ArrayList<>();
-
-				for (Mob entity : entities) {
-					if (entity instanceof Cow) {
-						convertNearby(entity, Items.BUCKET, Items.MILK_BUCKET);
-						if (entity instanceof MushroomCow) {
-							convertNearby(entity, Items.BOWL, Items.MUSHROOM_STEW);
-						}
-					} else if (entity instanceof Shearable shearable && shearable.readyForShearing()) {
-						shearables.add(entity);
-					}
-				}
-
-				Collections.shuffle(shearables);
-				int sheared = 0;
-
-				for (Mob entity : shearables) {
-					if (sheared > 4) {
-						break;
-					}
-
-					if (entity instanceof Shearable shearable) {
-						shearable.shear(SoundSource.BLOCKS);
-					}
-
-					++sheared;
-				}
+			switch (variant) {
+				case WILD -> HornItem.breakGrass(world, new ItemStack(BotaniaItems.grassHorn), pos, null);
+				case CANOPY -> HornItem.breakGrass(world, new ItemStack(BotaniaItems.leavesHorn), pos, null);
+				case GATHERING -> gatherProduce(world, pos);
 			}
 
 			for (int i = 0; i < 10; i++) {